requirements.txt
streamlitstreamlit
pandas
numpy
scikit-learn